+++ src/db/models/station_click_item_new.rs
@@ -1,0 +1,8 @@
+#[derive(Clone,Debug)]
+pub struct StationClickItemNew {
+    pub stationid: i32,
+    pub stationuuid: String,
+    pub ip: String,
+    pub clickuuid: String,
+    pub clicktimestamp: String,
+}