+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.2] 2020-01-11
+### Added
+- Sync for clicks
+- Api for clicks /format/clicks
+
+### Changed
+- Document "/format/url" endpoint as station click endpoint
+- Document returned structs "station", "checks"
+- Faster sync
+
+### Fixed
+- Clickcount calculation for each station
+- Error in /format/stations/changed endpoint
+
 ## [0.6.1] 2020-01-08
 ### Added
 - Prometheus docs